refactor: enum 클래스를 type 패키지로 이동 #1
